+++ app/javascript/packs/application.js
@@ -25,8 +25,6 @@ require("channels")
 // External imports
 import "bootstrap";
 
-// CSS
-import 'mapbox-gl/dist/mapbox-gl.css';
 
 // Internal imports, e.g:
 // import { initSelect2 } from '../components/init_select2';
@@ -40,40 +38,4 @@ document.addEventListener('turbolinks:load', () => {
   initMapbox();
 });
 
-
-
-
-import mapboxgl from 'mapbox-gl';
-
-const buildMap = (mapElement) => {
-  mapboxgl.accessToken = mapElement.dataset.mapboxApiKey;
-  return new mapboxgl.Map({
-    container: 'map',
-    style: 'mapbox://styles/mapbox/streets-v10'
-  });
-};
-
-const addMarkerToMap = (map, marker) => {
-    new mapboxgl.Marker()
-      .setLngLat([ marker.lng, marker.lat ])
-      .addTo(map);
-  };
-
-const fitMapToMarker = (map, marker) => {
-  const bounds = new mapboxgl.LngLatBounds();
-  mbounds.extend([ marker.lng, marker.lat ]);
-  map.fitBounds(bounds, { padding: 70, maxZoom: 15 });
-};
-
-const initMapbox = () => {
-  const mapElement = document.getElementById('map');
-  if (mapElement) {
-    const map = buildMap(mapElement);
-    const marker = JSON.parse(mapElement.dataset.marker);
-    addMarkerToMap(map, marker);
-    fitMapToMarker(map, marker);
-  }
-};
-
-export { initMapbox };
