+++ app/javascript/plugins/init_mapbox.js
@@ -1,0 +1,25 @@
+import mapboxgl from 'mapbox-gl';
+
+const initMapbox = () => {
+  const mapElement = document.getElementById('map');
+  const fitMapToMarker = (map, marker) => {
+    const bounds = new mapboxgl.LngLatBounds();
+    bounds.extend([ marker.lng, marker.lat ]);
+    map.fitBounds(bounds, { padding: 70, maxZoom: 15, duration: 0 });
+  };
+
+  if (mapElement) { // only build a map if there's a div#map to inject into
+    mapboxgl.accessToken = mapElement.dataset.mapboxApiKey;
+    const map = new mapboxgl.Map({
+      container: 'map',
+      style: 'mapbox://styles/cycode/cki05qkg40lep19qm4yimhfzc'
+    });
+    const marker = JSON.parse(mapElement.dataset.marker);
+    new mapboxgl.Marker()
+      .setLngLat([ marker.lng, marker.lat ])
+      .addTo(map);
+    fitMapToMarker(map, marker);
+  }
+};
+
+export { initMapbox };
