+++ app/views/goats/show.html.erb
@@ -29,12 +29,12 @@
 <div class="card">
    <%= "Description : #{@goat.description.capitalize}" %>
 </div>
-
+<br>
 <div id="map"
      style="width: 100%; height: 600px;"
      data-marker="<%= @marker.to_json %>"
      data-mapbox-api-key="<%= ENV['MAPBOX_API_KEY'] %>"></div>
-
+<br>
 <div class="btn-green btn-center">
   <%=link_to "Réservez-moi", new_goat_reservation_path(@goat) %>
 </div>
